% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cross2_dataset.R
\name{build_cross2_dataset}
\alias{build_cross2_dataset}
\title{build_cross2_dataset}
\usage{
build_cross2_dataset(
  geno_filename,
  pheno_filename,
  covar_column_names,
  pheno_column_names,
  genotypes = c("A", "H", "B"),
  alleles = c("A", "B"),
  na.strings = c("-1", "discrepancy", "NA"),
  comment.char = "#",
  ...
)
}
\arguments{
\item{geno_filename}{Path to csvs geno file}

\item{pheno_filename}{Path to csv pheno file}

\item{covar_column_names}{List of "phenotype" columns related to adjusting covariates.}

\item{pheno_column_names}{List of "phenotype" columns realted to main variables to study.}

\item{genotypes}{List of allowed genotype values of genotypes file}

\item{alleles}{List of homozygote alleles included, should be a subset of genotypes variable}

\item{na.strings}{List of strings associated to NA values}

\item{comment.char}{List or string of allowed comment chars}

\item{...}{Additional arguments passed to read.cross function}
}
\value{
This function returns an R/QTL Object of class "cross2". For details, see the \href{https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/QTL2} developer guide.
}
\description{
Generates cross2 dataset from basic geno and phenotypes name. Geno file should combine genenotype values and mapping (see to_csvs_geno) and phenotypes table should be a matrix-type object (users in rows, phenotypes in columns).
 This function is the entry point to QtlWrapper object.
}
\examples{
\dontrun{
  geno_data_matrix_filename = '/path/to/geno_file'
  geno_map_filename = '/path/to/geno_mapping_file'
  phenotype_filenam = '/path/to/phenotypes_file'

  genotypes_csvs  = to_csvs_geno(geno_map_filename, geno_data_matrix_filename)
  write_csvs_geno(genotypes_csvs, '/path_to_geno_csvs')

  chosen_genotypes    = c('A', 'H', 'B')
  chosen_alleles      = c('A','B')
  chosen_covars       = c('pheno_var1', 'pheno_var2', 'pheno_var3')
  chosen_phenotypes   = c('signature1', 'signature2', 'signature3')

  cross2 = systematicQTL::build_cross2_dataset(
    geno_filename = geno_filename,
    pheno_filename = pheno_filename,
    covar_column_names = chosen_covars,
    pheno_column_names = chosen_phenotypes,
    alleles = chosen_alleles
  )

}


}
