% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\name{build_genescan_plot}
\alias{build_genescan_plot}
\title{Draw genescan peaks plot}
\usage{
build_genescan_plot(object, phenotype, title)
}
\arguments{
\item{object}{A QTL object}

\item{phenotype}{Phenotype label to be displayed.}

\item{title}{Main title value}
}
\value{
plot
}
\description{
Draw genescan peaks plot
}
