% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\name{build_pdx_plot,QtlWrapper-method}
\alias{build_pdx_plot,QtlWrapper-method}
\title{Draw genescan peaks plot}
\usage{
\S4method{build_pdx_plot}{QtlWrapper}(object, chr, cm, phenotype, title, subtitle)
}
\arguments{
\item{object}{A QTL object}

\item{chr}{The SNP chromosome}

\item{cm}{The SNP distance in centimorgans}

\item{phenotype}{String with phenotype name to display over genotypes}

\item{title}{String with title plot}

\item{subtitle}{String with subtitle plot}
}
\value{
plot
}
\description{
Draw genescan peaks plot
}
