% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\name{find_peaks,QtlWrapper-method}
\alias{find_peaks,QtlWrapper-method}
\title{Finds the peaks passing lod values.
This method should be called after defining LOD values manually or empirically}
\usage{
\S4method{find_peaks}{QtlWrapper}(object)
}
\arguments{
\item{object}{A QTL object}
}
\value{
qtl
}
\description{
Finds the peaks passing lod values.
This method should be called after defining LOD values manually or empirically
}
