% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod,QtlWrapper-method}
\alias{lod,QtlWrapper-method}
\title{Retrieves real values LOD scores from a QTL wrapper object. If P.values have been defined (manually or empirically) also
returns significance stores}
\usage{
\S4method{lod}{QtlWrapper}(object)
}
\arguments{
\item{object}{An QTL object}
}
\value{
Data frame with genotype LOD scores
}
\description{
Retrieves real values LOD scores from a QTL wrapper object. If P.values have been defined (manually or empirically) also
returns significance stores
}
