% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\name{lod}
\alias{lod}
\title{Retrieves real values LOD scores from a QTL wrapper object. If P.values have been defined (manually or empirically) also
returns significance stores}
\usage{
lod(object)
}
\arguments{
\item{object}{An QTL object}
}
\value{
Data frame with genotype LOD scores
}
\description{
Retrieves real values LOD scores from a QTL wrapper object. If P.values have been defined (manually or empirically) also
returns significance stores
}
