% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\docType{class}
\name{QtlWrapper-class}
\alias{QtlWrapper-class}
\title{Qtl Object wrapper class.}
\description{
Qtl Object wrapper class.
}
\section{Slots}{

\describe{
\item{\code{qtl}}{QTL object}

\item{\code{phenotypes}}{Phenotype names}

\item{\code{map}}{Map object generated automatically at initialization}

\item{\code{pr}}{Calc_genoprob object generated automatically at initialization}

\item{\code{Xcovar}}{Matrix object with Xprobatilities}

\item{\code{genescan}}{A scan1 object generated at initialiation}

\item{\code{pvalues}}{Data frame with minimal LOD values thresholds by signature that should be considered as significant.}

\item{\code{peaks}}{Data Frame with computed peaks. Set at "find_peaks" method}
}}

