% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_empirical_significance_lod.R
\name{set_empirical_significance_lod}
\alias{set_empirical_significance_lod}
\title{Sets significance LOD threshold empirically using permutation tests}
\usage{
set_empirical_significance_lod(object, permutations, alpha, cores)
}
\arguments{
\item{object}{An QTL object}

\item{permutations}{Integer with the number of permutations to be executed}

\item{alpha}{Float value with the alpha value to be chosen, i.e. 0.05}

\item{cores}{Number of cores to use for run this permutation test}
}
\value{
qtl
}
\description{
Sets significance LOD threshold empirically using permutation tests
}
