% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtlWrapper.R
\name{set_significance_lod,QtlWrapper-method}
\alias{set_significance_lod,QtlWrapper-method}
\title{Sets significance LOD threshold for alpha < 0.05 manually.}
\usage{
\S4method{set_significance_lod}{QtlWrapper}(object, min_lod)
}
\arguments{
\item{object}{An QTL object}

\item{min_lod}{The minimal threshold value for all signatures}
}
\value{
qtl
}
\description{
Sets significance LOD threshold for alpha < 0.05 manually.
}
