% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_significance_lod.R
\name{set_significance_lod}
\alias{set_significance_lod}
\title{Sets significance LOD threshold for alpha < 0.05 manually.}
\usage{
set_significance_lod(object, min_lod)
}
\arguments{
\item{object}{An QTL object}

\item{min_lod}{The minimal threshold value for all signatures}
}
\value{
qtl
}
\description{
Sets significance LOD threshold for alpha < 0.05 manually.
}
