% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_csvs_geno.R
\name{to_csvs_geno}
\alias{to_csvs_geno}
\title{Combine markers map and  genotypes dataframe}
\usage{
to_csvs_geno(markers_map, genotypes)
}
\arguments{
\item{markers_map}{Dataset with markers}

\item{genotypes}{Dataset with phenotypes}
}
\value{
DataFrame for genotypes mapped in CSVS format
}
\description{
Combine markers map and  genotypes dataframe
}
\details{
Markers map file should contain 3 columns:

 - marker: Genomic marker Name (snp id for example)
 - chr: Chromosome
 - cm: Distance in centimorgans
 
Genotypes file must be a data frame containing at least a numeric "id" column and the markers in same order than markers map.
}
