% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvs_geno.R
\name{write_csvs_geno}
\alias{write_csvs_geno}
\title{Helper function to export csvs_geno dataset in correct format expected as input for buildCross2Dataset}
\usage{
write_csvs_geno(csvs_geno, filename)
}
\arguments{
\item{csvs_geno}{Dataframe (output of build_genos_df)}

\item{filename}{Where the data would be}
}
\description{
Helper function to export csvs_geno dataset in correct format expected as input for buildCross2Dataset
}
