% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cross2_dataset.R
\name{build_cross2_dataset}
\alias{build_cross2_dataset}
\title{build_cross2_dataset}
\usage{
build_cross2_dataset(
  geno_filename,
  pheno_filename,
  covar_column_names,
  pheno_column_names,
  genotypes = c("A", "H", "B"),
  alleles = c("A", "B"),
  na.strings = c("-1", "discrepancy", "NA"),
  comment.char = "#",
  ...
)
}
\arguments{
\item{geno_filename}{Path to csvs geno file}

\item{pheno_filename}{Path to csv pheno file}

\item{covar_column_names}{List of "phenotype" columns related to adjusting covariates.}

\item{pheno_column_names}{List of "phenotype" columns realted to main variables to study.}

\item{genotypes}{List of allowed genotype values of genotypes file}

\item{alleles}{List of homozygote alleles included, should be a subset of genotypes variable}

\item{na.strings}{List of strings associated to NA values}

\item{comment.char}{List or string of allowed comment chars}

\item{...}{Additional arguments passed to read.cross function}
}
\description{
Generates cross2 dataset from basic geno and phenotypes name. Geno file should combine genenotype values and mapping (see to_csvs_geno) and phenotypes table should be a matrix-type object (users in rows, phenotypes in columns)
}
